add @defgroup and @addtogroup to fix missing \ref
